models for rankings done
All tests succeed.
match_win_links and match_lose links are reference tables for matches.
If math with Id match_id is won, we look into the match_win_link table
for match_id and check the next_match_id to know which match the winner
of the match should play next