Integrated authentication with nifty-generators
Pages can now be blocked for unauthorised users by adding a before each
filter at the top of the controllers. I didnt do this yet, but the
infrastructure is there. I will now start adding these restrinctions to
the resources, and also test them.

For this commit all tests are working, except for 4 i comments out. They
were autogenerated by the nifty gem, but some of them seem to brake. the
behavior is there however. I put a TODO in the code.