Time in settings, time zones, css bug in games images
Added a start and end time/date in the settings, which will be used
in the info pages and the side bar.

Fixed some time zone inconsistencies by setting the time zone.

Fixed a bug that made the games images bigger than their divs.

Converted static page views to HAML.